GitHub Actions build refactor

+++ cpp/array_record_reader.cc
@@ -80,7 +80,7 @@ class ChunkOffset {
  public:
   virtual ~ChunkOffset() {}
   virtual uint64_t operator[](size_t idx) const = 0;
-  virtual size_t size() const = 0;
+  virtual uint64_t size() const = 0;
   bool empty() const { return size() == 0; }
 };
